ProjectBackendDev/requirements.txt
django
djangorestframework
psycopg2
drf-yasg
environs
mysqlclient